Add the database ID and bucket name to the status.
The patch fixes two issues:
1. Reconfiguring the sync daemon for a different bucket does not
   reset the Row ID.
2. If a drive with the DB failed, was replaced, and synced, the rows
   may be off. In that case we need to reset the Row ID and validate
   that all of the changes made their way into S3.

Since there are deployments that use the older scheme, we must ensure
to read it in correctly and write out the new scheme. Once everyone is
upgraded that code may be removed.

+++ s3_sync/sync_container.py
@@ -96,20 +96,48 @@ use = egg:swift#catch_errors
         self.s3_client.meta.events.unregister('before-call.s3.PutObject',
                                               conditionally_calculate_md5)
 
-    def load_status(self):
+    def load_status(self, db_id):
         if not os.path.exists(self._status_file):
-            return {}
+            return 0
         with open(self._status_file) as f:
             try:
-                return json.load(f)
+                status = json.load(f)
+                # First iteration did not include the bucket and DB ID
+                if 'last_row' in status:
+                    return status['last_row']
+                if db_id in status:
+                    entry = status[db_id]
+                    if entry['aws_bucket'] == self.aws_bucket:
+                        return entry['last_row']
+                    else:
+                        return 0
+                return 0
             except ValueError:
-                return {}
+                return 0
 
-    def save_status(self, state):
+    def save_status(self, row, db_id):
         if not os.path.exists(self._status_account_dir):
             os.mkdir(self._status_account_dir)
-        with open(self._status_file, 'w') as f:
-            json.dump(state, f)
+        if not os.path.exists(self._status_file):
+            with open(self._status_file, 'w') as f:
+                json.dump({db_id: dict(last_row=row,
+                                       aws_bucket=self.aws_bucket)}, f)
+                return
+
+        with open(self._status_file, 'rw') as f:
+            status = json.load(f)
+            # The first version did not include the DB ID and aws_bucket in the
+            # status entries
+            if 'last_row' in status:
+                status = {db_id: dict(last_row=row,
+                                      aws_bucket=self.aws_bucket)}
+            else:
+                status[db_id] = dict(last_row=row,
+                                     aws_bucket=self.aws_bucket)
+            f.seek(0)
+            json.dump(status, f)
+            f.truncate()
+            return
 
     def get_s3_name(self, key):
         concat_key = '%s/%s/%s' % (self.account, self.container, key)
